Cards that start with Tokens now have them
Several players requested that the Druid and the Necro should start with
their custom markers just like the Priest etc. So they now do, while I
was at it I added in the other markers for cards that start with them. I
think the logic is a bit spaghetti like but we can clean that up next
release.