Point the effect die roll to random.org
Exercise to get a better understanding of how the dice rolling code
works. Decided to leave it in.

Please look at what I did and fix my mistakes or poor form. If you don't
want the code in their long term we can remove it after we get done
playing with it.