My take on supporting multiple boards #67
taking what I learned and this is what I came up with...is there a
better way?

One thought is in the OnTableLoad Section of actions.py is to pop a menu
with board selections. The code would be cleaner from that perspective,
but the menu will support a large number of boards.