Probabilities Module Started
A new python file (it can be easily merged later if necessary, I just
thought it would be best to keep new things separate for now) that
contains the code for computing the expected damage and chance to kill a
creature given a particular amount of dice and armor. It isn't
integrated into the octgn module at all, but the computations work
perfectly. Also, unlike my earlier combat simulator, it uses
combinatorics rather than test runs to get its numbers, so it is
consistent and perfectly accurate.