Expected Damage Module + Fix to getStat()
The Expected Damage Module now works! Granted, there are a lot of side
cases that don't work, but if you take the average creature or
conjuration, target it, and roll dice, you will see the probable outcome
for each number of dice. Also, I fixed getStat(), which was not working
because the delimiter should be ', ' rather than ',' (because the xml
file containing the cards uses a comma and a space to separate each
different stat).

Anyway, there is quite a bit more work to be done to make the EDM
airtight and user-friendly, but the basics work now.