Added attacker/defender detection to dice stuff
This is a pretty exciting change. With the new functionality, if you
target one creature and then execute "Attack Targeted Creature" on
another creature, the game will prompt you to select one of the latter's
attacks to use against the former. For now, this only works between
creatures and conjurations under normal circumstances (and is still
missing a bit of code), but it works well enough that you should be able
to get some idea of what I am doing.

I have great plans for this...