Integrated TAB into untargeted die roll
- TAB is now the all-purpose dice rolling button. If nothing is targeted
when you press it, you get a generic dice menu if the cursor is not over
a card, and the attack menu (sans probability calculations) for a card
if your cursor is hovering over that card. If you press it while one
card is targeted, you get the same things (depending on where the cursor
is), but with battle calculations included.

- Fixes a bug that was preventing the untargeted (Attacker =
None,Defender = None) die roll from working properly.