Added function to evaluate attack dice bonuses
When targeted attacks are declared, the game checks for modifiers and
adjusts the numbers of dice rolled for the attack accordingly.
Obviously, it currently covers very few dice buffs and no debuffs, so it
will need some fleshing out.

Also corrected an XML formatting mistake. Assumed that the standard
format places a space between the end of the text and the start of the
attachTraits granted.