#126 Dice direct input menu remembers last entry
With the new dice buttons, selecting a die roll should be faster. On
occasion, players may still need to roll a different number of dice,
however. Ideally, this will not happen very often, but when it does, the
player might need to repeatedly enter this number for subsequent dice
rolls (for example, if your mage is rolling 11 dice per turn for some
reason not picked up by the attack detector). To make things faster in
this case, the default entry for the dice input menu is now the last
number you typed into it.