#140 - Implemented basic attachment costs
Now upon attaching a face-down enchantment to a card, the game will ask
if you want to pay for it, then offer you a choice of how much mana to
pay. The basic functionality seems to work, but discounts are not
completely integrated, and I think the text could be improved.