#156 - Found workaround to fix problem
#156 - I still have no idea why #156 occurs, but some experimentation
revealed that it only seems to happen when table.create is called
directly from the onGameStart event or the onTableLoad event, or any
function associated with those events (and possibly in other events as
well, though I haven't tested it). So I moved this part of the code to
the deck loading event; the first time any player loads a deck, these
areas will be placed as usual. This might result in some bugs if players
try to use these buttons and areas, but we already prohibit dice-rolling
before deck-loading, so I don't think this is a huge issue. Anyway, it
works as a temporary measure until we can figure out how to get around
the root problem (or OCTGN fixes it, if it is a bug).

This commit also has unfinished work on one aspect of #144; normally I
would have waited until it worked right, but I felt that fixing #156
takes priority.