#144 - Completed basic damage receipt code
This completes the unfinished section of code started in
https://github.com/sIKE23/Mage-Wars/commit/ac3e5032109b96c2dc97eafe6224dfb2d812bc3c,
though there is still much to be done with this particular feature
(completes in the sense that it now works as intended, though it ignores
a lot of special cases).

Now, when you use TAB+target to order a creature (or attack spell) to
attack a specific object, the game will calculate how much damage and
what conditions (if any) the attack inflicts, based on your roll and the
properties of the attacker, attack, and defender. Then, the controller
of the object that was attacked is asked whether they would like to
apply the specified amount of damage. If they say yes, that amount of
damage is automatically applied to that object, along with any
conditions that were inflicted.