#157 - Fixed, and future-proofed
Fixed bug #157 by having the discard action also detach the card. Also
altered the attachCards code so that if attached cards end up in some
non-table pile, they should not exhibit this behavior (though we won't
know if this works until that happens). This should save us grief of a
similar sort in the future.