#144 - getAttackList now attributes attack spells
When your mage declares an attack, the game will search for all of your
face-up attack spells that are:
a. Not attached
b. Attached to your mage
c. Attached to an object that is not capable of making an attack itself
(such as elemental wand)

and add these to the list of attacks. If your familiar or spawnpoint
declares an attack, the game will add all face-up attack cards attached
to that familar or spawnpoint to its attack list.

Also, with the exception of equipment with the autonomous trait or cards
with damage barriers, you can no longer order a piece of equipment or an
attack spell to declare an attack; you have to have the mage (or an
appropriate familiar) do it. This is good because it ensures that the
game attributes the attack to the correct source.