#144 - Default traits and compound traits
#144 - Now the game makes some assumptions about the traits that a card
has based on its type; for instance, creatures are assumed to be living
and corporeal.

#144 - Traits that imply other traits now grant those traits. For
example, an incorporeal creature is automatically recognized as having
the Nonliving, Burnproof, and Uncontainable traits.

#144 - Having a guard marker grants the counterstrike trait to a
creature's melee attacks (though, not, now that I think about it, only
its quick attacks; I'll have to fix that in the next commit). Similarly,
Sleep and Stun grant the incapacitated trait.

#144 - Changed the way immunity is formatted to fix bug where immunities
were overriding each other.

#144 - Fixed a bug that was preventing the battle calculator from
working properly with walls.