#144 - Significant improvement of effectDetection
#144 - When you declare an attack, and the target is immune to one or
more of the effects, those effects do not show up on the menu and are
not considered when rolling. This takes into consideration elemental
immunities, the unmovable trait, and the burnproof trait. So, for
example, the Snatch effect roll for a Thornlasher is not displayed when
declaring an attack against an Iron Golem (or any conjuration). If the
attack has multiple effects of different types, only those of a type to
which the target is not immune will be shown; for instance, Devil's
Trident used against a poison immune creature would read
Burn(25%),Burn(33%) (Which is perhaps less than ideal, but I still think
it's pretty good).

#144 - Trait autodetection now considers rooted creatures to have the
Unmovable trait, and to not have the flying trait.

#137 - Fixed XML mistakes on Adramelech Warlock and Blood Demon