#162 - Should fix validation error
Gemini on the news groups identified the issue. It was with the code
that you updated the last time around. Combined with a change in how the
deck builder now returns null values, it used to be 0 and is now
"empty". The result was that you can't covert an empty string to an
integer. We poked the following into the code below line 463:

if (ret == String.Empty)
ret = "0";

and that has resolved the issue. The Mage and Magestat cards do not have
most of the properties used in this section of the code.....

Long term I am going to look at having someone re-factor this code from
the one big method that it is to smaller ones to make it easier to
maintain.