#144 - Implemented damage ring bonuses
#144 - Damage specific Melee/Range bonuses from
Fireshaper/Dawnbreaker/Lightning ring now implemented correctly.

#137, 146 - I removed the bracketed traits from the damage rings,
because they interfered with parsing. We do not need generic recognition
of type-specific damage boosters because there are so few cards with
this property that the amount of code for reading the traits correctly
is roughly comparable to the amount of code for writing 3 exceptions
(which I have done). Actually, it might be even greater.