#144 - MAJOR update: Defenses implemented!
#144 - Now defenses work! Well, there is still quite a bit to be done,
but creatures can at least use their innate defense against attacks.

#144 - The game now remembers what has happened so far this round, or
this turn. This is only used for keeping track of defenses so far, but
will be useful later on when memory effects like Melee +X (which only
applies to the first attack in a given round)

#144 - Major restructuring; I am moving toward structuring attacks
around the seven steps of an attack, each of which leads into the next.
This should clean things up a lot.

#137 - Python does not seem to like parsing ascii infinity, so I changed
it to inf (for defenses)