#144 - Implemented OR traits
#144 - I decided that the simplest way to implement OR traits on attacks
(for instance, Doublestrike OR Sweeping) was to have them split up into
multiple attacks. So, for instance, if you order Sir Corazin to make an
attack, you will see 3 options; his quick attack, his full attack with
sweeping, and his full attack with doublestrike. This allows players to
decide which trait they want to use when they declare the attack.

#137 - Fixed lack of OR capitalization on Ludwig.