#166 - Implemented text/map loader
#166 - The game is now capable of loading and reading .txt files.
Currently, this is only implemented as proof of concept; loading a file
will just make it display as a string in the messages box. But this
proves that it works. Loading a file gives an array (formatted as a set
of columns), which we can use for the automatic setup.

Also implemented an interface for loading the maps. I figured out how to
read all the txt files in our chosen directory, so instead of having the
players type in the name of the file that they want (prone to errors),
it simply presents them with a list of txt files located in the maps
directory instead, form which they may choose their desired map. This
should be much more user friendly.

To test this feature out, open the group actions menu and select 'Load
Map'.