#166 - Added multiple layer functionality to loader
#166 - The map loader now can read text files with multiple layers.
Formatting notes:
- Each layer of the map should have the name of the layer surrounded by
# characters directly below it. For instance, to mark a layer as the Map
layer, write #Map# below that layer. See my examples for clarification.
- Each layer is stored as an array (a list of columns) in a dictionary
object, with the key set to the name of the layer. So if you want to
get, for instance, the Objects array, you would take the object returned
from importArray (which is a dictionary) and query the name of the layer
that you want (e.g. <<< mapArray =
importArray(<filename>).get('Objects',<default value>).