#166 - MAJOR update: map loader now functional
#166 - Though there are still a few issues to fix (like spiked pit trap)
and things on which to decide (like the coordinates of the center of the
map), the map loader is now fully functional. I made a demonstration map
that you can load to see the full potential (loading multiple objects,
different zones, etc.). I could describe it, or you could just sync it,
design a few maps, and try it out.