#144 - Exciting update: Zone detection implemented!
#144 - The game now recognizes zones, and knows what cards are in or
bordering them. It turns out you can't store objects in global
variables, so I defined the map and zones as dictionaries, which works
just as well.

Currently, the only function of zones is that objects will snap to their
appropriate locations (non-walls snap to the inside of the zone, and
walls snap to the appropriate border).

#166 - Now that zone detection works, I can implement mage starting
positions (but have not done so yet).