#144 - Range enforcement implemented
#144 - Attack legality based on range is now enforced. Only attacks for
which the attacker is within range of the defender show up on the attack
menu. For instance, a Royal Archer can only use her knife attack if her
target is in the same zone, and can only use her bow if the target is
1-2 zones away. Also applies to ranged healing abilities, like the
Asyran Cleric's. Abilities without a specified range (like the Gray
Angel's) can target things, anywhere.

#144 - Fixed bug that was preventing attack spell ranges from being
computed properly.

#144 - Fixed bug that was causing attack colors to display incorrectly.