#144 - Daze and Disable implemented
#144 - If a creature is dazed, its defense rolls are now reduced by 2
per daze marker, and it must successfully roll a 7 or above to attack,
upon which an appropriate message will be displayed (depending on the
outcome of the daze roll). Daze is not autoremoved at the end of the
turn, since I currently have no way to know when a player is done acting
(I have previously tried to tie similarly timed features to CTRL+RIGHT,
but it didn't work for some reason, and I don't know if everybody uses
CTRL+RIGHT).

#144 - Disabled equipment now does not give bonuses to the controlling
mage (or at least, does not give combat bonuses; I don't know if you
have implemented disabling upkeep bonuses, but if not you can easily do
so by routing your functions through computeTraits(card)).