#144 - Attack costs implemented
#144 - Now, when a player attacks with an attack that has a cost, they
are prompted to pay the mana from their supply.

#144 - If the Battle Calculator is enabled, players will not be prompted
to pay mana when revealing face-down attack spells (since this prompt
will be given when they attempt to actually use the spell). The first
time (and only the first time a player is surprised by this (expecting
attack spells to prompt for reveal costs), a message will appear to
inform that player of the change.

#144 - Effects that take away (and do not allow for the return of) the
flying trait, in particular guarding, now apply (so guarding flying
creatures can be attacked in melee by non-flyers).

#137 - Misc. XML fixes