#151 - Bound spells correct attributed to source
#151 - Now, when a face-up spell is cast, its source is correctly
attributed. If it is bound to a familiar or spawnpoint, that will be
given in the notification as the source of the spell (e.g. "Huginn,
Raven Familiar casts Drain Power, with a printed casting cost of 16").
If unbound, or bound to a wand, it will treat the mage as the source of
the spell by default.

#178 - More death messages