#191 - Problem solved
#191 - Fixed the bug preventing phases from working properly. The
problem was that a phase marker was being created for each player, and
these markers were both treated as the phase marker, leading to de-sync
issues (since There Can Be Only One). Fixed it by moving it to
initiative marker creation, which already happens only once. Phases
should work normally now (but we should test extensively to be sure).

#137 - Minor XML fix

#183 - Began construction of a function to automate most of the labor of
parsing XML code to make the new spell dictionary. Will drastically
reduce the formatting work once complete (we will just need to define
the functions). This is for Q2, but the code does not actually interact
with the octgn module at all, so it won't affect the Q1 release even if
it is included (and I would rather not branch yet).