#191 - Fixed "Invalid Marker Count" messages
#191 - Fixed "Invalid Marker Count" error messages. The problem was that
each resolveChanneling function was attempting to resolveChanneling for
both players rather than just the local player.

#151,#188 - Should not prompt to pay for enchantment when ordering game
to cast an enchantment on an illegal target (I have not tested this)

Misc. - resolveChanneling does not consider Mana Crystal/Flower or
Harmonize, the reasoning being that players generally adjust their
channeling value, and we don't want an invisible channeling bonus on top
of that. It would be better to auto-adjust their channeling value itself
when these cards are played/discarded, but for now I think we should
just let them take care of this themselves.