#122 - Channeling +1 cards now increase a mages Channeling stats
Also when the card(s) are discarded the Channeling stat will be
decreased.

Notes: Harmonize must be attached to the mage when it is revealed to
effect the mages stats and the same when it is discarded. If it is to
the side of the board at this point it will neither increase nor
decrease a mages stats.

When we move to the Spell Dictionary method I am hoping attaching and
detaching cards will trigger additions and subtractions such as this.

#191 - Had some new Phase Markers made for the re-factor of the Phase
function.  Though I have pushed this to Q2,, thought that we could use
some of the markers now.