#144 - Fixed reported bug with Bear Strength
#144 - Fixed the following reported bug:

>If you have Bear Strength hidden on a creature or mage when it starts
the attack sequence, you are prompted to see if you want to reveal it,
if you say yes, you are prompted to pay for the enchantment, it is
revealed but the melee bonus is not applied to this attack sequence. It
is applied to the next one though.

It affected more than just Bear strength.

#207 - Redid parts of code using deepcopy to make sure it works. It
does! Hooray!