#221 - Fixes Attachments and Bindings
The change to the OnMoveCards event was the replacement of isScriptMove
parameter with the faceup parameter. Since the former isScriptMove
parameter always return false your if not isScriptMove: block was always
true (and therefore no reason to test) so I removed it from the function
had shifted the child indents under it back one indentation level. This
appears to of fixed the problem. You can find the updated package on
Dropbox, I am going to have Schwenkgott test it for me and if it doesn't
break anything I am going to push out the next release.

The good news is the abiltiy now exists to see if a card is faceup/down
when you move it.