#144 - Fixed JBM Wounded Prey ability reassignment
#144 - The JBM's wounded prey marker is returned to her when the target
dies. Note that this may not work properly if more than 1 JBM is in
prey; unfortunately, markers cannot be associated with players. I know
how to fix this problem, but the amount of work that it will take is not
worth the payoff, so I recommend that we just leave this bug for Q2,
when the machinery will be in place.

#144 - Fixed a bug with Wounded prey; it now gives the bonus to the
correct creatures (animals and the mage), rather than any creature.