#255 - Moved Color Selection to new Color Dictionary
This is the first part to get the Control Marker work done. I did a good
bit of time prototyping this in another game module. It still needs a
good amount of testing but I think the work is solid, and I think
stylistically close to how you would do it. If you have a better idea
for the the layout of the Dictionary I am more than welcome to
suggestions.