#255 - There is a single group action to place a marker on an Orb now.
It also handles removing a marker and setting it to be "neutral". I also
moved the logic from addToken back to the calling function. Its not
resuable and makes much better sense when not split in half.

If you get to it before I do, we should be to test if the defender is an
"Orb" and in the Damage and Effects phase or Interimstep phase of
attacks.py have it call addControlMarker(defenderCard)