#166 - Added in a prompt once the Map is Selected
to Let Players know what map is being played and what the win conditions
are. I had issues with formatting sentences with Map names some of them
really needed a "the" and on others the "the" made no sense what so
ever. So I created a text function to provide me a custom sentence for
map. The custom text strings are to follow, I have outsourced that task
for the moment

#166 - fixed a map name with an extra space in its name

#259 - backed out the changes for this, I will need @ACG8 to look at and
fix this logic
#201 - Added in Game End logic resulting from a win by the most V'Tar