#166 - Guard Tokens are only added to Orb Guardians if they are Scenario Guardian
and they are in a Zone with an Orb. I know they can't be pushed or
teleported out of the zone with Orb and this is a bit of overkill, but I
don't like the unexpected.

#166 - Domination Markers as Cards now have the Persist tag and will be
sent to the Obliterate pile upon discard. I decided that this would
probably work best in-case a player accidentally discarded one of these
cards.