Changed gitignore file to ignore decks/images
I got tired of copypasting my mage wars code each time there was an
update, so I decided to map the Mage-Wars directory to the same location
that OCTGN uses for the module. I altered the gitignore file so that I
can store images and decks in the same directory without them being
uploaded with each commit.

#95 - Also deleted the superfluous timer.py file (which was moved to
actions.py previously.