#213 - Created code-friendly targeting format
Created a new card property that holds a targeting format that is
code-friendly to greatly simplify determining legal targets for spells.
Put a guide to the syntax in main.py and wrote a formatter in the xml
mass editor (#286) to make inserting the new property into existing xml
files as easy as possible.

Also created a property for storing buffs, with similar
syntax.(#216,#215)