#294 - Began integration of TVM into attach system
The TVM now (partially) works with the attachment system; the hideous,
exception-ridden code from canAttach has been replaced with a much
simpler one. It won't work with all attachments yet, and isn't
integrated with the system that computes all of a creatures traits, but
it can handle some of the more common target types. I also mass-inserted
a cTargets property into all the xml data using #286 to support this
module.