#176 - Can now declare attacks through TAB
A player can select "declare attack" as a TAB action option. This allows
them to create a targeting arrow/attack menu by left-clicking on a
target.

#249 - Cleaned up the arrow-targeting code by moving the bulk of it out
of the event. Also discovered a major bug (not sure how it got there)
that would cause players to pay twice for an attack.